pyproject.toml
[project]
name = "try-medgenma"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.7.0",
    "pillow>=11.2.1",
    "torch>=2.7.1",
    "transformers>=4.52.4",
]
